requirements.txt
# A股分析系统 - 依赖清单
# 安装命令: pip install -r requirements.txt

# ========== 核心数据 ==========
akshare>=1.14.0          # 主数据源（免费A股数据）
baostock>=0.8.8          # 历史数据备用
pandas>=2.0.0            # 数据处理
numpy>=1.24.0            # 数值计算

# ========== HTTP请求 ==========
requests>=2.31.0         # HTTP请求
aiohttp>=3.9.0           # 异步HTTP（可选）

# ========== 数据存储 ==========
pyyaml>=6.0               # 读取automation.yaml配置
sqlite3                   # 内置，无需安装（历史数据存储）

# ========== AI分析（免费额度）==========
groq>=0.4.0               # Groq API（免费LLM）
openai>=1.12.0            # 兼容OpenAI接口的API

# ========== 定时任务 ==========
# APScheduler>=6.0.0       # 本地定时（可选）
# croniter>=2.0.0          # Cron表达式解析

# ========== 微信推送 ==========
# serverchan-sdk>=1.0.0   # Server酱SDK（可选，直接requests调用）

# ========== 图表 ==========
matplotlib>=3.8.0         # 绑图
plotly>=5.18.0            # 交互图表

# ========== 日志 ==========
loguru>=0.7.0             # 日志库

# ========== 开发工具 ==========
python-dotenv>=1.0.0      # 环境变量
tqdm>=4.66.0              # 进度条
